Tweaks
- Move `mergeOptions` function to separate function
- Other tweaks to improve code
